day_64/requirements.txt
Flask-SQLAlchemy==2.5.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
WTForms==2.3.3
requests~=2.25.1
python-dotenv~=0.19.0